fix: DashboardLayout, hydration error 이슈 수정
