+++ src/app/auth/layout.tsx
@@ -9,12 +9,8 @@ export default function AuthLayout({
   children: React.ReactNode
 }) {
   return (
-    <html lang="en">
-      <body>
         <main className="flex h-screen justify-center items-center">
           {children}
         </main>
-      </body>
-    </html>
   )
 }
