+++ src/components/layout/ModalHandler.tsx
@@ -1,0 +1,12 @@
+'use client'
+
+import { useModalStore } from '@/store/useModalStore'
+import dynamic from 'next/dynamic'
+
+const MyPage = dynamic(() => import('./Mypage'), { ssr: false })
+
+export default function ModalHandler() {
+  const { isModalOpen, closeModal } = useModalStore()
+
+  return <MyPage isOpen={isModalOpen} onClose={closeModal} />
+}
